// fron/cliente/src/app/models/Edm01.ts
export class Edm01casillas{
    _id?: number;
    edm01casilla1: string;
    edm01casilla2: string;
    edm01casilla3: string;
    edm01casilla4: string;
    edm01casilla5: string;
    edm01casilla6: string;
    edm01casilla7: string;
    edm01casilla8: string;
    edm01casilla9: string;
    edm01casilla10: string;
    edm01casilla11: string;
    edm01casilla12: string;
    edm01casilla13: string;
    edm01casilla14: string;
    edm01casilla15: string;
    edm01casilla16:string;
    edm01casilla17: string;
    edm01casilla18: string;
    edm01casilla19: string;
    edm01casilla20: string;
    edm01casilla21: string;
    edm01casilla22: string;
    edm01casilla23: string;
    edm01casilla24: string;
    edm01casilla25: string;
    edm01casilla26: string;
    edm01casilla27: string;
    

    constructor(edm01casilla1:string, edm01casilla2:string,edm01casilla3:string,edm01casilla4:string,edm01casilla5:string ,edm01casilla6: string,
       edm01casilla7: string,edm01casilla8: string,edm01casilla9: string, edm01casilla10: string,edm01casilla11: string,edm01casilla12: string,edm01casilla13: string,edm01casilla14: string,
        edm01casilla15: string,edm01casilla16: string,edm01casilla17: string,edm01casilla18: string,edm01casilla19: string,edm01casilla20: string,edm01casilla21: string,
        edm01casilla22: string,edm01casilla23: string, edm01casilla24: string,edm01casilla25: string,edm01casilla26: string,edm01casilla27: string,
       ){
        this.edm01casilla1=edm01casilla1;
        this.edm01casilla2=edm01casilla2;
        this.edm01casilla3=edm01casilla3;
        this.edm01casilla4=edm01casilla4;
        this.edm01casilla5=edm01casilla5;
        this.edm01casilla6=edm01casilla6;
        this.edm01casilla7=edm01casilla7;
        this.edm01casilla8=edm01casilla8;
        this.edm01casilla9=edm01casilla9;
        this.edm01casilla10=edm01casilla10;
        this.edm01casilla11=edm01casilla11;
        this.edm01casilla12=edm01casilla12;
        this.edm01casilla13=edm01casilla13;
        this.edm01casilla14=edm01casilla14;
        this.edm01casilla15=edm01casilla15;
        this.edm01casilla16=edm01casilla16;
        this.edm01casilla17=edm01casilla17;
        this.edm01casilla18=edm01casilla18;
        this.edm01casilla19=edm01casilla19;
        this.edm01casilla20=edm01casilla20;
        this.edm01casilla21=edm01casilla21;
        this.edm01casilla22=edm01casilla22;
        this.edm01casilla23=edm01casilla23;
        this.edm01casilla24=edm01casilla24;
        this.edm01casilla25=edm01casilla25;
        this.edm01casilla26=edm01casilla26;
        this.edm01casilla27=edm01casilla27;

    }

}